Update name to DateTimePickerControl
