--- packages/js/components/src/date-time-picker/index.tsx
@@ -1,1 +1,0 @@
-export * from './date-time-picker';
